+++ src/Symfony/Component/Config/Resource/FileResource.php
@@ -17,8 +17,10 @@ namespace Symfony\Component\Config\Resource;
  * The resource can be a file or a directory.
  *
  * @author Fabien Potencier <fabien@symfony.com>
+ *
+ * @final since Symfony 4.3
  */
-class FileResource implements SelfCheckingResourceInterface, \Serializable
+class FileResource implements SelfCheckingResourceInterface
 {
     /**
      * @var string|false
@@ -61,21 +63,5 @@ class FileResource implements SelfCheckingResourceInterface, \Serializable
     public function isFresh($timestamp)
     {
         return false !== ($filemtime = @filemtime($this->resource)) && $filemtime <= $timestamp;
-    }
-
-    /**
-     * @internal
-     */
-    public function serialize()
-    {
-        return serialize($this->resource);
-    }
-
-    /**
-     * @internal
-     */
-    public function unserialize($serialized)
-    {
-        $this->resource = unserialize($serialized);
     }
 }