+++ src/Symfony/Component/DependencyInjection/Config/ContainerParametersResource.php
@@ -17,8 +17,10 @@ use Symfony\Component\Config\Resource\ResourceInterface;
  * Tracks container parameters.
  *
  * @author Maxime Steinhausser <maxime.steinhausser@gmail.com>
+ *
+ * @final since Symfony 4.3
  */
-class ContainerParametersResource implements ResourceInterface, \Serializable
+class ContainerParametersResource implements ResourceInterface
 {
     private $parameters;
 
@@ -36,22 +38,6 @@ class ContainerParametersResource implements ResourceInterface, \Serializable
     public function __toString()
     {
         return 'container_parameters_'.md5(serialize($this->parameters));
-    }
-
-    /**
-     * @internal
-     */
-    public function serialize()
-    {
-        return serialize($this->parameters);
-    }
-
-    /**
-     * @internal
-     */
-    public function unserialize($serialized)
-    {
-        $this->parameters = unserialize($serialized);
     }
 
     /**